Added fake actions for mcu stats, reboot, shutdown

+++ src/store/api/mcu/mcuReboot.js
@@ -2,7 +2,8 @@ const { exec } = require('child_process')
 
 module.exports = ({ define }) => {
   define('reboot', async (payload, { knex, errors, utils }) => {
-    exec('sudo shutdown -r now')
+    if (process.env.NODE_ENV === 'production') return exec('sudo shutdown -r now');
+    return;
   }, {
     auth: true
   })
